✨ Feat: 상세페이지 csr로 재전환 및 컨테이너 컴포넌트를 page로 이동
