+++ src/app/(after-login)/epigrams/[id]/page.tsx
@@ -1,37 +1,86 @@
-import { notFound } from 'next/navigation';
-import { dehydrate, HydrationBoundary } from '@tanstack/react-query';
-import { isAxiosError } from 'axios';
-import { axiosServerInstance } from '@/utils/axios';
-import { getQueryClient } from '@/utils/getQueryClient';
+'use client';
+
 import Inner from '@/components/Inner';
-import EpigramDetail from '../_components/EpigramDetail';
 import EpigramComments from '../_components/EpigramComments';
+import { notFound, useParams, useRouter } from 'next/navigation';
+import { useSession } from 'next-auth/react';
+import { useEpigram } from '@/apis/epigram/epigram.queries';
+import { useState } from 'react';
+import DeleteModal from '@/components/DeleteModal';
+import {
+  DeatailFooter,
+  DetailContent,
+  DetailHeader,
+  DetailLoading,
+  DetailWrapper,
+} from '../_components/detail';
 
-export default async function Page({ params }: { params: Promise<{ id: string }> }) {
-  const id = Number((await params).id);
-  const queryClient = getQueryClient();
+export default function Page() {
+  const params = useParams<{ id: string }>();
+  const router = useRouter();
+  const { data: session } = useSession();
+  const [isDeleteConfirmModalOpen, setIsDeleteConfirmModalOpen] = useState(false);
 
-  try {
-    await queryClient.fetchQuery({
-      queryKey: ['epigrams', id],
-      queryFn: async () => {
-        const response = await axiosServerInstance.get(`/epigrams/${id}`);
-        return response.data;
-      },
-    });
-  } catch (error) {
-    if (isAxiosError(error) && error.response?.status === 404) {
-      notFound();
+  const id = Number(params.id);
+  const { details, remove, like, disLike } = useEpigram(id);
+
+  if (details.isLoading) return <DetailLoading />;
+  if (details.isError) throw new Error('게시물을 가져오지 못했습니다.');
+  if (!details.data) return notFound();
+
+  const { tags, content, author, writerId, referenceTitle, referenceUrl, likeCount, isLiked } =
+    details.data;
+  const isOwner = writerId === session?.user.id;
+
+  const handleEdit = () => {
+    if (!isOwner) {
+      return alert('작성자만 수정이 가능합니다.');
     }
 
-    throw new Error('게시물을 가져오는데 문제가 생겼습니다.');
-  }
+    router.push(`/epigrams/${id}/edit`);
+  };
+
+  const handleRemove = () => {
+    if (!isOwner) {
+      return alert('작성자만 삭제가 가능합니다.');
+    }
+
+    remove.mutate();
+  };
+
+  const handleRemoveConfirm = () => {
+    setIsDeleteConfirmModalOpen(true);
+  };
+
+  const handleLike = () => {
+    return !isLiked ? like.mutate() : disLike.mutate();
+  };
 
   return (
     <div className='bg-bg flex min-h-dvh flex-col gap-8 md:gap-10 lg:gap-12'>
-      <HydrationBoundary state={dehydrate(queryClient)}>
-        <EpigramDetail id={id} />
-      </HydrationBoundary>
+      <DetailWrapper>
+        <DetailHeader
+          tags={tags}
+          isOwner={isOwner}
+          onEdit={handleEdit}
+          onRemove={handleRemoveConfirm}
+        />
+        <DetailContent content={content} author={author} />
+        <DeatailFooter
+          referenceTitle={referenceTitle}
+          referenceUrl={referenceUrl}
+          isLiked={isLiked}
+          likeCount={likeCount}
+          onLike={handleLike}
+        />
+      </DetailWrapper>
+      <DeleteModal
+        isOpen={isDeleteConfirmModalOpen}
+        type='post'
+        onClose={() => setIsDeleteConfirmModalOpen(false)}
+        onDelete={handleRemove}
+        isSubmitting={remove.isPending}
+      />
 
       <Inner>
         <EpigramComments id={id} />
